Redirect child process stdin only when instructed to do so
There are cases where direct input to the child process is
necessary. For example we might want to run the child process with
root privileges in which case we invoke it with sudo and provide our
password. This change ensures that we redirect the child process stdin
only if we use the --stdin argument.